Replace the vtree cost with a width score
The old `cost` was a function of where clauses meet: max load cubed, the
cross-subtree clause product, and a clause-scope term. A chain vtree keeps
every one of those small, so the numbering-only baselines (`linear`,
`reverse-linear`) ranked above every decomposition-based construction on
`cost` while compiling to diagrams two to three orders of magnitude larger.

`vtree_cost` now scores the separators. For each internal node the clauses
crossing it are counted three ways — variables below the node that occur
above it, variables above it that occur below, and the crossing clauses
themselves — and the smallest count is the node's width `w`. The score is
`log2 Σ 2^w` over the nodes with `w > 0`, in bits. On 46 reduced MCC
formulas × 24 vtrees each, compiled under every vtree, its rank correlation
with the measured peak diagram size is 0.70 against 0.28 for the old
formula, and no baseline ranks first on any of them (the old formula: all).

`VtreeScores::cost` becomes `f64`; everything that read it as `u64`
(selection tie-break, the portfolio's adopt rule, the root sweep and the
flowcutter/hybrid/goatd pick sites) follows. `components.json` bumps to
`vitri-components-v2`. The claim that `2^ctx` bounds the intermediate
diagram is removed: it does not (one inside variable under `a ∨ c` and
`¬a ∨ d` has three subfunctions), and the score is documented as what it
is measured to be. The fixture pins the three per-node tables and
`cost == 3.0`; the README example is regenerated (its pick moves to
`goatd` on the new tie-break).

+++ src/decompose/goatd/tests/end_to_end.rs
@@ -155,11 +155,11 @@ fn multi_seed_produces_valid_tds() {
 use crate::decompose::best::select_first_min;
 use crate::decompose::goatd::refined_select_key;
 
-type Cand = (&'static str, u32, usize, u64);
+type Cand = (&'static str, u32, usize, f64);
 
 /// The scored path's key, mirroring the tuple `best_vtree_over_schedule`
 /// builds inline: lowest width wins, ties broken by cost then bag size.
-fn scored_select_key(width: u32, bagsize: usize, cost: u64) -> (u64, u64, u64) {
+fn scored_select_key(width: u32, bagsize: usize, cost: f64) -> (u64, f64, u64) {
     (width as u64, cost, bagsize as u64)
 }
 
@@ -167,8 +167,8 @@ fn scored_select_key(width: u32, bagsize: usize, cost: u64) -> (u64, u64, u64) {
 fn refined_key_orders_by_width_then_bagsize() {
     // Same width; A must prefer the smaller bagsize regardless of cost.
     let cands: Vec<Cand> = vec![
-        ("small_bag_high_qual", 5, 10, 100),
-        ("big_bag_low_qual", 5, 20, 50),
+        ("small_bag_high_qual", 5, 10, 100.0),
+        ("big_bag_low_qual", 5, 20, 50.0),
     ];
     let winner = select_first_min(cands.iter().copied(), |&(_, w, b, _)| {
         refined_select_key(w, b)
@@ -182,8 +182,8 @@ fn scored_key_orders_by_width_then_cost() {
     // Same input as the refined test; the scored path breaks the tie on
     // cost (2nd field), so it picks the OTHER candidate.
     let cands: Vec<Cand> = vec![
-        ("small_bag_high_qual", 5, 10, 100),
-        ("big_bag_low_qual", 5, 20, 50),
+        ("small_bag_high_qual", 5, 10, 100.0),
+        ("big_bag_low_qual", 5, 20, 50.0),
     ];
     let winner = select_first_min(cands.iter().copied(), |&(_, w, b, q)| {
         scored_select_key(w, b, q)
@@ -196,7 +196,7 @@ fn scored_key_orders_by_width_then_cost() {
 fn select_first_min_breaks_ties_toward_first() {
     // Two candidates with identical refined keys: the FIRST listed wins,
     // matching `min_by_key`'s tie-break rule.
-    let cands: Vec<Cand> = vec![("first", 7, 15, 1), ("second", 7, 15, 2)];
+    let cands: Vec<Cand> = vec![("first", 7, 15, 1.0), ("second", 7, 15, 2.0)];
     let winner = select_first_min(cands.iter().copied(), |&(_, w, b, _)| {
         refined_select_key(w, b)
     })
